# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Facility {
  id: number;
  name: string;
  shortName: string;
  description: string;
  priceCents: number;
  accent: string;
  capacity: number;
  /** Number of independently bookable courts within this facility */
  courts: number;
}

export type AvailabilitySlotStatus = typeof AvailabilitySlotStatus[keyof typeof AvailabilitySlotStatus];


export const AvailabilitySlotStatus = {
  available: 'available',
  reserved: 'reserved',
  blocked: 'blocked',
} as const;

export interface AvailabilitySlot {
  facilityId: number;
  court: number;
  date: string;
  startTime: string;
  endTime: string;
  status: AvailabilitySlotStatus;
  /** @nullable */
  reservationId: number | null;
}

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 10 */
  phone: string;
  email: string;
}

export type Lead = LeadInput & {
  id: number;
  createdAt: string;
};

export type ReservationInputPaymentMethod = typeof ReservationInputPaymentMethod[keyof typeof ReservationInputPaymentMethod];


export const ReservationInputPaymentMethod = {
  pix: 'pix',
  card: 'card',
} as const;

export interface ReservationInput {
  facilityId: number;
  /** Which specific court to reserve, for facilities with multiple courts (defaults to 1) */
  court?: number;
  date: string;
  startTime: string;
  endTime: string;
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  customerPhone: string;
  customerEmail: string;
  paymentMethod: ReservationInputPaymentMethod;
}

export type ReservationStatus = typeof ReservationStatus[keyof typeof ReservationStatus];


export const ReservationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export type Reservation = ReservationInput & ({
  id: number;
  facilityName: string;
  priceCents: number;
  status: ReservationStatus;
  /** @nullable */
  bookingCode?: string | null;
  createdAt: string;
});

export interface BookingItem {
  facilityId: number;
  court?: number;
  date: string;
  startTime: string;
  endTime: string;
}

export type BookingInputPaymentMethod = typeof BookingInputPaymentMethod[keyof typeof BookingInputPaymentMethod];


export const BookingInputPaymentMethod = {
  pix: 'pix',
  card: 'card',
} as const;

export interface BookingInput {
  /**
     * @minItems 1
     * @maxItems 40
     */
  items: BookingItem[];
  /** @minLength 2 */
  customerName: string;
  /** @minLength 10 */
  customerPhone: string;
  customerEmail: string;
  paymentMethod: BookingInputPaymentMethod;
}

export interface BookingPix {
  /** Pix copy-and-paste (BR Code) with the amount already fixed */
  payload: string;
  amountCents: number;
}

export type BookingPaymentMethod = typeof BookingPaymentMethod[keyof typeof BookingPaymentMethod];


export const BookingPaymentMethod = {
  pix: 'pix',
  card: 'card',
} as const;

export interface Booking {
  code: string;
  totalCents: number;
  paymentMethod: BookingPaymentMethod;
  reservations: Reservation[];
  pix: BookingPix | null;
}

export type ReservationUpdateStatus = typeof ReservationUpdateStatus[keyof typeof ReservationUpdateStatus];


export const ReservationUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface ReservationUpdate {
  status: ReservationUpdateStatus;
}

export interface AdminSummary {
  reservationsToday: number;
  confirmedThisMonth: number;
  expectedRevenueCents: number;
  activeCustomers: number;
}

export interface BlockedSlotInput {
  facilityId: number;
  court?: number;
  date: string;
  startTime: string;
  endTime: string;
  reason?: string;
}

export type BlockedSlot = BlockedSlotInput & {
  id: number;
  createdAt: string;
};

export type DateQueryParameter = string;

export type ListAvailabilityParams = {
date: DateQueryParameter;
facilityId?: number;
};

export type ListReservationsParams = {
from?: string;
status?: ListReservationsStatus;
};

export type ListReservationsStatus = typeof ListReservationsStatus[keyof typeof ListReservationsStatus];


export const ListReservationsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;