# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentMethod } from './bookingPaymentMethod';
import type { BookingPix } from './bookingPix';
import type { Reservation } from './reservation';

export interface Booking {
  code: string;
  totalCents: number;
  paymentMethod: BookingPaymentMethod;
  reservations: Reservation[];
  pix: BookingPix | null;
}